+++ teku/src/main/java/tech/pegasys/teku/services/BeaconNodeServiceController.java
@@ -24,7 +24,6 @@ import tech.pegasys.teku.services.executionlayer.ExecutionLayerService;
 import tech.pegasys.teku.services.powchain.PowchainService;
 import tech.pegasys.teku.validator.client.ValidatorClientService;
 import tech.pegasys.teku.validator.client.slashingriskactions.DoppelgangerDetectionShutDown;
-import tech.pegasys.teku.validator.client.slashingriskactions.SlashedValidatorAlert;
 import tech.pegasys.teku.validator.client.slashingriskactions.SlashedValidatorShutDown;
 import tech.pegasys.teku.validator.client.slashingriskactions.SlashingRiskDetectionAction;
 
@@ -55,16 +54,18 @@ public class BeaconNodeServiceController extends ServiceController {
             tekuConfig.discovery().isDiscoveryEnabled()));
     powchainService(tekuConfig, serviceConfig, maybeExecutionWeb3jClientProvider)
         .ifPresent(services::add);
-    final SlashingRiskDetectionAction validatorSlashedAction =
+
+    final Optional<SlashingRiskDetectionAction> maybeValidatorSlashedAction =
         tekuConfig.validatorClient().getValidatorConfig().isStopWhenValidatorSlashedEnabled()
-            ? new SlashedValidatorShutDown()
-            : new SlashedValidatorAlert();
+            ? Optional.of(new SlashedValidatorShutDown())
+            : Optional.empty();
+
     services.add(
         ValidatorClientService.create(
             serviceConfig,
             tekuConfig.validatorClient(),
             new DoppelgangerDetectionShutDown(),
-            validatorSlashedAction));
+            maybeValidatorSlashedAction));
   }
 
   private Optional<PowchainService> powchainService(