+++ teku/src/main/java/tech/pegasys/teku/services/ValidatorNodeServiceController.java
@@ -13,11 +13,11 @@
 
 package tech.pegasys.teku.services;
 
+import java.util.Optional;
 import tech.pegasys.teku.config.TekuConfiguration;
 import tech.pegasys.teku.service.serviceutils.ServiceConfig;
 import tech.pegasys.teku.validator.client.ValidatorClientService;
 import tech.pegasys.teku.validator.client.slashingriskactions.DoppelgangerDetectionShutDown;
-import tech.pegasys.teku.validator.client.slashingriskactions.SlashedValidatorAlert;
 import tech.pegasys.teku.validator.client.slashingriskactions.SlashedValidatorShutDown;
 import tech.pegasys.teku.validator.client.slashingriskactions.SlashingRiskDetectionAction;
 
@@ -25,15 +25,15 @@ public class ValidatorNodeServiceController extends ServiceController {
 
   public ValidatorNodeServiceController(
       final TekuConfiguration tekuConfig, final ServiceConfig serviceConfig) {
-    final SlashingRiskDetectionAction validatorSlashedAction =
+    final Optional<SlashingRiskDetectionAction> maybeValidatorSlashedAction =
         tekuConfig.validatorClient().getValidatorConfig().isStopWhenValidatorSlashedEnabled()
-            ? new SlashedValidatorShutDown()
-            : new SlashedValidatorAlert();
+            ? Optional.of(new SlashedValidatorShutDown())
+            : Optional.empty();
     this.services.add(
         ValidatorClientService.create(
             serviceConfig,
             tekuConfig.validatorClient(),
             new DoppelgangerDetectionShutDown(),
-            validatorSlashedAction));
+            maybeValidatorSlashedAction));
   }
 }